# server/docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  controller.handleEventCreate.eventDataType:
    properties:
      name:
        example: Klickrausch
        type: string
      username:
        example: 2
        type: integer
    required:
    - name
    - username
    type: object
  controller.handleUserCreate.userDataType:
    properties:
      isAdmin:
        example: false
        type: boolean
      name:
        example: Max Mustermann
        type: string
      password:
        example: catsAreAwesome
        type: string
      username:
        example: mmustermann
        type: string
    required:
    - name
    - password
    - username
    type: object
  controller.handleUserUpdate.userDataType:
    properties:
      isAdmin:
        example: false
        type: boolean
      name:
        example: Max Mustermann
        type: string
      passworChangeRequired:
        example: false
        type: boolean
      password:
        example: catsAreAwesome
        type: string
      username:
        example: mmustermann
        type: string
    required:
    - name
    - username
    type: object
  controller.publicUser:
    properties:
      createdAt:
        example: "2023-05-29T15:34:14.198515266+02:00"
        type: string
      id:
        example: 17
        type: integer
      isAdmin:
        example: false
        type: boolean
      name:
        example: Mark Maximus
        type: string
      passwordChangeRequired:
        example: false
        type: boolean
      updatedAt:
        example: "2023-05-30T15:12:35.463734634+02:00"
        type: string
      username:
        example: maximum
        type: string
    type: object
info:
  contact: {}
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Yotei API
  version: v1
paths:
  /event:
    get:
      description: Get all events based on permissions
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Get all events
      tags:
      - Event
    post:
      description: Every user can create a new event
      parameters:
      - description: Data for new event
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.handleEventCreate.eventDataType'
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Create a new event
      tags:
      - Event
  /event/{accessKey}:
    delete:
      description: Only the creator and all administrators can delete an event
      parameters:
      - description: Event access key
        in: path
        name: accessKey
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete an event
      tags:
      - Event
    get:
      description: Can be used by anyone who haves the access key
      parameters:
      - description: Event access key
        in: path
        name: accessKey
        required: true
        type: string
      responses:
        "200":
          description: Event
          schema:
            properties:
              event:
                type: string
            type: object
        "400":
          description: Error message
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Event not found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get one event
      tags:
      - Event
    patch:
      description: Everyone can update a participant except there is an user linked
        to it
      parameters:
      - description: Event access key
        in: path
        name: accessKey
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Update a participant of the current event
      tags:
      - Event
    post:
      description: Everybody can use this route, if the user is logged in, the participant
        will be linked to the user
      parameters:
      - description: Event access key
        in: path
        name: accessKey
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Add a participant to an existing event
      tags:
      - Event
    put:
      description: Only the creator and all administrators can update an event
      parameters:
      - description: Event access key
        in: path
        name: accessKey
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Update an event
      tags:
      - Event
  /login:
    post:
      description: Login an existing user with a username and password
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Login user
      tags:
      - Authentication
  /logout:
    post:
      description: Logout the current user
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Logout user
      tags:
      - Authentication
  /user:
    get:
      description: Only works for administrators
      responses:
        "200":
          description: List of all users
          schema:
            properties:
              users:
                items:
                  $ref: '#/definitions/controller.publicUser'
                type: array
            type: object
        "403":
          description: Only admins can access the user list
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get all users
      tags:
      - User
    post:
      description: Only works for administrators
      parameters:
      - description: Data for new user
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.handleUserCreate.userDataType'
      responses:
        "200":
          description: User ID of newly created user
          schema:
            properties:
              userId:
                type: integer
            type: object
        "400":
          description: Error message
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create a new user
      tags:
      - User
  /user/{userId}:
    delete:
      description: Only works for administrators and the user itself
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "200":
          description: User ID of deleted user
          schema:
            properties:
              userId:
                type: integer
            type: object
        "400":
          description: Error message
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Only admins and the user itself can delete a user
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Delete an user
      tags:
      - User
    get:
      description: Only works for administrators and the user itself
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "200":
          description: User
          schema:
            properties:
              user:
                $ref: '#/definitions/controller.publicUser'
            type: object
        "400":
          description: Error message
          schema:
            properties:
              error:
                type: string
            type: object
        "403":
          description: Only admins and the user itself can access a user
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: User not found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get one user
      tags:
      - User
    put:
      description: Only works for administrators and the user itself
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: integer
      - description: Data for updated user
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.handleUserUpdate.userDataType'
      responses:
        "200":
          description: User ID of updated user
          schema:
            properties:
              userId:
                type: integer
            type: object
        "400":
          description: Error message
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update an user
      tags:
      - User
produces:
- application/json
swagger: "2.0"